+++ src/main/java/org/openlmis/fulfillment/i18n/MessageKeys.java
@@ -28,6 +28,7 @@ public abstract class MessageKeys {
   private static final String USER = "user";
   private static final String ORDER = "order";
   private static final String ORDER_RETRY = "orderRetry";
+  private static final String ORDER_UPDATE = "orderUpdate";
   private static final String ORDER_FILE_TEMPLATE = "fileTemplate";
   private static final String REFERENCE_DATA = "referenceData";
   private static final String REPORTING = "reporting";
@@ -89,6 +90,9 @@ public abstract class MessageKeys {
 
   public static final String ORDER_RETRY_INVALID_STATUS =
       join(ERROR_PREFIX, ORDER_RETRY, INVALID_STATUS);
+
+  public static final String ORDER_UPDATE_INVALID_STATUS =
+      join(ERROR_PREFIX, ORDER_UPDATE, INVALID_STATUS);
 
   public static final String ERROR_ORDER_FILE_TEMPLATE_CREATION =
       join(ERROR_PREFIX, ORDER_FILE_TEMPLATE, CREATION);